switching to tests instead of plugins after adding cache feature
