+++ helpers/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+	cache: require('./cache.js'),
+	functions: require('./functions.js'),
+	regions: require('./regions.js')
+};
