+++ helpers/regions.js
@@ -1,0 +1,21 @@
+var regions = [
+	'us-east-1',
+	'us-west-1',
+	'us-west-2',
+	'ap-northeast-1',
+	'ap-southeast-1',
+	'ap-southeast-2',
+	'eu-central-1',
+	'eu-west-1',
+	'sa-east-1'
+];
+
+module.exports = {
+	all: regions,
+	cloudtrail: regions,
+	ec2: regions,
+	elb: regions,
+	kms: regions,
+	vpc: regions,
+	rds: regions
+};
